Upgrade gm dependency
The specified range "^1.18.1" allowed the version 1.19.0 to be
installed. This version has a bug which caused certain images
to trigger an ECONNRESET issue.

A test has been added which fails with gm@1.19.0 installed and
passes with subsequent fixed versions (1.20 and upwards). The
version range in package.json has been bumped to prevent the bad
version being installed.